pyproject.toml
[project]
name = "rahgiri-bot"
version = "0.3.0"
description = "A Telegram bot for tracking parcels via the Iran Post Tracking System."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fake-useragent>=2.0.3",
    "playwright>=1.49.1",
    "pydantic-settings>=2.7.1",
    "python-telegram-bot>=21.9",
]

[dependency-groups]
dev = [
    "isort>=5.13.2",
    "mypy>=1.14.1",
    "ruff>=0.8.4",
]

[tool.mypy]
namespace_packages = false
files = ["bot/**/*.py"]
check_untyped_defs = true
disable_error_code = ["empty-body"]
disallow_untyped_defs = false
disallow_subclassing_any = false
strict = true

[tool.ruff]
include = ["*.py", "*.pyi", "**/pyproject.toml"]
line-length = 130

[tool.isort]
atomic = true
profile = "black"
known_first_party = ["black", "blib2to3", "blackd", "_black_version"]

